Some speed-up in cost computation. All costs end up as integers now, also target costs. Given that costs are of the order of 1 million, rounding errors can be neglected. Further speedup is needed.
git-svn-id: https://mary.opendfki.de/repos/trunk@120 953a6561-930b-0410-b2a6-db37d1b2ae63

+++ java/de/dfki/lt/mary/unitselection/TargetCostFunction.java
@@ -47,7 +47,7 @@ public interface TargetCostFunction
      * @param unit
      * @return a non-negative number; smaller values mean better fit, i.e. smaller cost.
      */
-    public float cost(Target target, Unit unit);
+    public int cost(Target target, Unit unit);
     
     /**
      * Set the features of the cost function
